consensus: Extract harness helpers into support modules
